+++ wpull/processor/rule.py
@@ -1,11 +1,20 @@
 '''Fetching rules.'''
+import logging
 import random
 
 from trollius import From, Return
 import trollius
+
+from wpull.backport.logging import BraceMessage as __
 from wpull.hook import HookableMixin, HookDisconnected, Actions, HookStop
-from wpull.item import Status
+from wpull.item import Status, LinkType
 from wpull.errors import DNSNotFound, ServerError, ConnectionRefused
+from wpull.scraper.css import CSSScraper
+from wpull.scraper.html import HTMLScraper
+import wpull.url
+
+
+_logger = logging.getLogger(__name__)
 
 
 class FetchRule(HookableMixin):
@@ -145,7 +154,7 @@ class FetchRule(HookableMixin):
 
         return verdict, reason
 
-    def check_ftp_request(self, url_info, url_record):
+    def check_generic_request(self, url_info, url_record):
         '''Check URL filters and scripting hook.
 
         Returns:
@@ -159,6 +168,8 @@ class FetchRule(HookableMixin):
 
         return verdict, reason
 
+    check_ftp_request = check_generic_request
+
 
 class ResultRule(HookableMixin):
     '''Decide on the results of a fetch.
@@ -168,6 +179,8 @@ class ResultRule(HookableMixin):
             to be a permanent error.
         retry_dns_error: If True, don't consider a DNS resolution error to be
             permanent error.
+        waiter (:class:`.waiter.Waiter`): The Waiter.
+        statistics (:class:`.stats.Statistics`): The Statistics.
     '''
     def __init__(self, retry_connrefused=False, retry_dns_error=False,
                  waiter=None, statistics=None):
@@ -179,9 +192,11 @@ class ResultRule(HookableMixin):
 
         self.register_hook(
             'wait_time', 'handle_response', 'handle_pre_response',
-            'handle_error')
+            'handle_error',
+        )
 
     def handle_pre_response(self, request, response, url_item):
+        '''Process a response that is starting.'''
         action = self.consult_pre_response_hook(
             request, response, url_item.url_record)
 
@@ -298,6 +313,7 @@ class ResultRule(HookableMixin):
         return action
 
     def get_wait_time(self):
+        '''Return the wait time in seconds between requests.'''
         seconds = self._waiter.get()
         try:
             return self.call_hook('wait_time', seconds)
@@ -305,6 +321,7 @@ class ResultRule(HookableMixin):
             return seconds
 
     def consult_pre_response_hook(self, request, response, url_record):
+        '''Return scripting action when a response begins.'''
         try:
             return self.call_hook(
                 'handle_pre_response', request, response, url_record
@@ -313,6 +330,7 @@ class ResultRule(HookableMixin):
             return Actions.NORMAL
 
     def consult_response_hook(self, request, response, url_record):
+        '''Return scripting action when a response ends.'''
         try:
             return self.call_hook(
                 'handle_response', request, response, url_record
@@ -321,7 +339,97 @@ class ResultRule(HookableMixin):
             return Actions.NORMAL
 
     def consult_error_hook(self, request, url_record, error):
+        '''Return scripting action when an error occured.'''
         try:
             return self.call_hook('handle_error', request, url_record, error)
         except HookDisconnected:
             return Actions.NORMAL
+
+
+class ProcessingRule(HookableMixin):
+    '''Document processing rules.
+
+    Args:
+        fetch_rule (FetchRule): The FetchRule instance.
+        document_scraper (:class:`.scaper.DemuxDocumentScraper`): The document
+            scraper.
+    '''
+    def __init__(self, fetch_rule, document_scraper=None):
+        super().__init__()
+
+        self._fetch_rule = fetch_rule
+        self._document_scraper = document_scraper
+
+        self.register_hook('scrape_document')
+
+    parse_url = staticmethod(wpull.url.parse_url_or_log)
+
+    def scrape_document(self, request, response, url_item):
+        '''Process document for links.'''
+        if not self._document_scraper:
+            return
+
+        demux_info = self._document_scraper.scrape_info(request, response)
+
+        num_inline_urls = 0
+        num_linked_urls = 0
+
+        for scraper, scrape_info in demux_info.items():
+            new_inline, new_linked = self._process_scrape_info(
+                scraper, scrape_info, url_item
+            )
+            num_inline_urls += new_inline
+            num_linked_urls += new_linked
+
+        _logger.debug(__('Candidate URLs: inline={0} linked={1}',
+                         num_inline_urls, num_linked_urls
+        ))
+
+        try:
+            self.call_hook(
+                'scrape_document', request, response, url_item
+            )
+        except HookDisconnected:
+            pass
+
+    def _process_scrape_info(self, scraper, scrape_info, url_item):
+        '''Collect the URLs from the scrape info dict.'''
+        if not scrape_info:
+            return 0, 0
+
+        if isinstance(scraper, CSSScraper):
+            link_type = LinkType.css
+        elif isinstance(scraper, HTMLScraper):
+            link_type = LinkType.html
+        else:
+            link_type = None
+
+        inline_urls = scrape_info['inline_urls']
+        linked_urls = scrape_info['linked_urls']
+
+        inline_url_infos = set()
+        linked_url_infos = set()
+
+        for url in inline_urls:
+            url_info = self.parse_url(url)
+            if url_info:
+                url_record = url_item.child_url_record(
+                    url_info, inline=True
+                )
+                if self._fetch_rule.check_generic_request(url_info, url_record)[0]:
+                    inline_url_infos.add(url_info)
+
+        for url in linked_urls:
+            url_info = self.parse_url(url)
+            if url_info:
+                url_record = url_item.child_url_record(
+                    url_info, link_type=link_type
+                )
+                if self._fetch_rule.check_generic_request(url_info, url_record)[0]:
+                    linked_url_infos.add(url_info)
+
+        url_item.add_inline_url_infos(inline_url_infos)
+        url_item.add_linked_url_infos(
+            linked_url_infos, link_type=link_type)
+
+        return len(inline_url_infos), len(linked_url_infos)